requirements.txt
black>=24.1.0
Django>=4.2.0
django-cors-headers>=4.0.0
django-upgrade>=1.15.0
djangorestframework>=3.14.0
factory-boy>=3.3.0
Faker>=19.0.0
gunicorn>=21.0.0
isort>=5.13.0
Pillow>=10.0.0
pre-commit>=3.6.0
psycopg2-binary>=2.9.0
pylint>=3.2.0
pylint-django>=2.5.3
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-django>=4.7.0
pytest-factoryboy>=2.6.1
python-decouple>=3.8
whitenoise>=6.5.0
